refactor(engineer_loop/verification_actions): split tests to sibling tests_verification_actions.rs (#1266)
Inline test module in verification_actions.rs is moved into a sibling
tests_verification_actions.rs using the existing engineer_loop convention.

- src/engineer_loop/verification_actions.rs: 414 → 285 LOC (under 400-line rule)
- src/engineer_loop/tests_verification_actions.rs: new (~127 LOC)
- src/engineer_loop/mod.rs: register mod tests_verification_actions
- super::super::types::* → super::types::* in extracted body (sibling level)

Pure refactor: zero logic changes, all 9 tests pass.

Closes 1/70 of #1266 (after #1285-#1291).
